requirements.txt
scrapy
scrapy-playwright
EbookLib
requests
pillow